+++ CHANGELOG.md
@@ -22,7 +22,7 @@ This release is a big one with the introduction of a `Wgit::DSL` and Javascript 
 - `Wgit::Document` methods: `#at_xpath`, `#at_css` - which call nokogiri underneath.
 - `Wgit::Document#extract` method to perform one off content extractions.
 - `Wgit::Indexer#index_urls` method which can index several urls in one call.
-- `Wgit::Url` methods: `#to_user`, `#to_password`.
+- `Wgit::Url` methods: `#to_user`, `#to_password`, `#to_sub_domain`.
 ### Changed/Removed
 - Breaking change: Moved all `Wgit.index*` convienence methods into `Wgit::DSL`.
 - Breaking change: Removed `Wgit::Url#normalise`, use `#normalize` instead.